@@ -18,90 +18,91 @@
   "author": "Luis Rodrigues <luis.rodrigues@seedstarslabs.com>",
   "license": "MIT",
   "dependencies": {
-    "autoprefixer": "6.4.0",
-    "babel-core": "6.13.2",
+    "autoprefixer": "6.5.1",
+    "babel-core": "6.17.0",
     "babel-loader": "6.2.5",
     "babel-plugin-react-display-name": "2.0.0",
     "babel-plugin-react-transform": "2.0.2",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-plugin-undeclared-variables-check": "6.8.0",
-    "babel-polyfill": "6.13.0",
-    "babel-preset-es2015": "6.13.2",
-    "babel-preset-es2015-loose": "7.0.0",
-    "babel-preset-react": "6.11.1",
-    "babel-preset-stage-1": "6.13.0",
+    "babel-polyfill": "6.16.0",
+    "babel-preset-es2015": "6.16.0",
+    "babel-preset-es2015-loose": "8.0.0",
+    "babel-preset-react": "6.16.0",
+    "babel-preset-stage-1": "6.16.0",
     "babel-runtime": "6.11.6",
     "body-parser": "1.15.2",
+    "bootstrap-loader": "1.2.1",
+    "bootstrap-sass": "3.3.7",
     "classnames": "2.2.5",
-    "css-loader": "0.23.1",
+    "css-loader": "0.25.0",
     "dotenv": "2.0.0",
     "errorhandler": "1.4.3",
-    "es6-promise": "3.2.1",
+    "es6-promise": "4.0.5",
     "exports-loader": "0.6.3",
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.9.0",
-    "foundation-sites": "6.2.3",
     "font-awesome": "4.6.3",
     "font-awesome-webpack": "0.0.4",
-    "history": "3.0.0",
+    "history": "4.3.0",
     "html-webpack-plugin": "2.22.0",
     "imports-loader": "0.6.5",
     "isomorphic-fetch": "2.2.1",
-    "jquery": "3.1.0",
-    "jsonwebtoken": "7.1.9",
-    "jwt-decode": "2.1.0  ",
+    "jquery": "3.1.1",
     "less": "2.7.1",
     "less-loader": "2.2.3",
     "method-override": "2.3.6",
-    "node-sass": "3.8.0",
+    "node-sass": "3.10.1",
     "postcss-import": "8.1.0",
-    "postcss-loader": "0.10.1",
-    "react": "15.3.1",
-    "react-dom": "15.3.1",
+    "postcss-loader": "0.13.0",
+    "react": "15.3.2",
+    "react-dom": "15.3.2",
     "react-mixin": "3.0.5",
     "react-redux": "4.4.5",
-    "react-router": "2.7.0",
-    "react-router-redux": "4.0.5",
-    "redux": "3.5.2",
+    "react-router": "2.8.1",
+    "react-router-redux": "4.0.6",
+    "redux": "3.6.0",
     "redux-thunk": "2.1.0",
     "resolve-url-loader": "1.6.0",
-    "sass-loader": "4.0.0",
+    "sass-loader": "4.0.2",
     "style-loader": "0.13.1",
+    "tcomb-form": "0.9.10",
     "url-loader": "0.5.7",
-    "webpack": "2.1.0-beta.8",
+    "webpack": "2.1.0-beta.15",
     "webpack-merge": "0.14.1",
     "whatwg-fetch": "1.0.0",
-    "yargs": "5.0.0"
+    "yargs": "6.0.0"
   },
   "devDependencies": {
-    "babel-eslint": "6.1.2",
+    "babel-eslint": "7.0.0",
     "chai": "3.5.0",
-    "chai-as-promised": "5.3.0",
+    "chai-as-promised": "6.0.0",
+    "clean-webpack-plugin": "0.1.13",
     "enzyme": "2.4.1",
-    "eslint": "3.3.1",
-    "eslint-config-airbnb": "10.0.1",
-    "eslint-plugin-import": "1.13.0",
-    "eslint-plugin-jsx-a11y": "2.1.0",
-    "eslint-plugin-react": "6.1.2",
+    "eslint": "3.7.1",
+    "eslint-config-airbnb": "12.0.0",
+    "eslint-plugin-import": "2.0.1",
+    "eslint-plugin-jsx-a11y": "2.2.3",
+    "eslint-plugin-react": "6.4.1",
     "expect": "1.20.2",
-    "ignore-styles": "4.0.0",
+    "ignore-styles": "5.0.01",
     "istanbul": "1.0.0-alpha.2",
-    "jsdom": "9.4.2",
-    "mocha": "3.0.2",
+    "jsdom": "9.6.0",
+    "mocha": "3.1.2",
     "mocha-junit-reporter": "1.12.0",
-    "nock": "8.0.0",
-    "react-addons-test-utils": "15.3.1",
+    "nock": "8.1.0",
+    "react-addons-test-utils": "15.3.2",
     "react-transform-catch-errors": "1.0.2",
     "react-transform-hmr": "1.0.4",
-    "redbox-react": "1.3.0",
+    "redbox-react": "1.3.1",
     "redux-devtools": "3.3.1",
     "redux-devtools-dock-monitor": "1.1.1",
     "redux-devtools-log-monitor": "1.0.11",
-    "redux-logger": "2.6.1",
-    "redux-mock-store": "1.1.4",
-    "sass-lint": "1.9.0",
-    "sinon": "1.17.5",
-    "webpack-dev-middleware": "1.6.1",
-    "webpack-hot-middleware": "2.12.2"
+    "redux-logger": "2.7.0",
+    "redux-mock-store": "1.2.1",
+    "sass-lint": "1.9.1",
+    "sinon": "1.17.6",
+    "webpack-dev-middleware": "1.8.4",
+    "webpack-hot-middleware": "2.13.0"
   }
 }
