@@ -3,9 +3,26 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import classNames from 'classnames';
 import { push } from 'react-router-redux';
+import t from 'tcomb-form';
 
 import * as actionCreators from '../../actions/auth';
 
+const Form = t.form.Form;
+
+const Login = t.struct({
+    email: t.String,
+    password: t.String
+});
+
+const LoginFormOptions = {
+    auto: 'placeholders',
+    help: <i>Hint: a@a.com / qw</i>,
+    fields: {
+        password: {
+            type: 'password',
+        }
+    }
+};
 
 class LoginView extends React.Component {
 
@@ -14,8 +31,12 @@ class LoginView extends React.Component {
         isAuthenticated: React.PropTypes.bool.isRequired,
         isAuthenticating: React.PropTypes.bool.isRequired,
         statusText: React.PropTypes.string,
-        actions: React.PropTypes.object.isRequired,
-        location: React.PropTypes.object // this comes from react-router, not required
+        actions: React.PropTypes.shape({
+            authLoginUser: React.PropTypes.func.isRequired,
+        }).isRequired,
+        location: React.PropTypes.shape({
+            query: React.PropTypes.object.isRequired
+        })
     };
 
     constructor(props) {
@@ -23,8 +44,10 @@ class LoginView extends React.Component {
 
         const redirectRoute = this.props.location ? this.props.location.query.next || '/' : '/';
         this.state = {
-            email: '',
-            password: '',
+            formValues: {
+                email: '',
+                password: ''
+            },
             redirectTo: redirectRoute
         };
     }
@@ -35,29 +58,30 @@ class LoginView extends React.Component {
         }
     }
 
-    login = (e) => {
-        e.preventDefault();
-        this.props.actions.authLoginUser(this.state.email, this.state.password, this.state.redirectTo);
+    onFormChange = (value) => {
+        this.setState({ formValues: value });
     };
 
-    handleInputChange = (e, state) => {
-        this.setState({
-            [state]: e.currentTarget.value
-        });
+    login = (e) => {
+        e.preventDefault();
+        const value = this.loginForm.getValue();
+        if (value) {
+            this.props.actions.authLoginUser(value.email, value.password, this.state.redirectTo);
+        }
     };
 
     render() {
         let statusText = null;
         if (this.props.statusText) {
             const statusTextClassNames = classNames({
-                alert: true,
-                alert__error: this.props.statusText.indexOf('Authentication Error') === 0,
-                alert__success: this.props.statusText.indexOf('Authentication Error') !== 0
+                'alert': true,
+                'alert-danger': this.props.statusText.indexOf('Authentication Error') === 0,
+                'alert-success': this.props.statusText.indexOf('Authentication Error') !== 0
             });
 
             statusText = (
                 <div className="row">
-                    <div className="small-12 columns">
+                    <div className="col-sm-12">
                         <div className={statusTextClassNames}>
                             {this.props.statusText}
                         </div>
@@ -67,54 +91,24 @@ class LoginView extends React.Component {
         }
 
         return (
-            <div className="container">
-                <div className="row margin-top-large">
-                    <div className="small-6 small-centered columns">
-                        <div className="row">
-                            <div className="small-12 columns">
-                                <h1>Log in to view protected content!</h1>
-                            </div>
-                        </div>
-                        <div className="row margin-top-medium">
-                            <div className="small-12 columns">
-                                <p>Hint: a@a.com / qw</p>
-                            </div>
-                        </div>
-
-                        {statusText}
-
-                        <form>
-                            <div className="row margin-top-small">
-                                <div className="small-12 columns">
-                                    <input type="text"
-                                           className="form-control input-lg"
-                                           placeholder="Email"
-                                           onChange={(e) => { this.handleInputChange(e, 'email'); }}
-                                    />
-                                </div>
-                            </div>
-                            <div className="row margin-top-small">
-                                <div className="small-12 columns">
-                                    <input type="password"
-                                           className="form-control input-lg"
-                                           placeholder="Password"
-                                           onChange={(e) => { this.handleInputChange(e, 'password'); }}
-                                    />
-                                </div>
-                            </div>
-                            <div className="row margin-top-small">
-                                <div className="small-12 columns">
-                                    <button type="submit"
-                                            className="button button-medium float-right"
-                                            disabled={this.props.isAuthenticating}
-                                            onClick={this.login}
-                                    >
-                                        Submit
-                                    </button>
-                                </div>
-                            </div>
-                        </form>
-                    </div>
+            <div className="container login">
+                <h1 className="text-center">Login</h1>
+                <div className="login-container margin-top-medium">
+                    {statusText}
+                    <form onSubmit={this.login}>
+                        <Form ref={(ref) => { this.loginForm = ref; }}
+                              type={Login}
+                              options={LoginFormOptions}
+                              value={this.state.formValues}
+                              onChange={this.onFormChange}
+                        />
+                        <button disabled={this.props.isAuthenticating}
+                                type="submit"
+                                className="btn btn-default btn-block"
+                        >
+                            Submit
+                        </button>
+                    </form>
                 </div>
             </div>
         );
