@@ -7,8 +7,8 @@ import routes from '../../routes';
 export default class Root extends React.Component {
 
     static propTypes = {
-        store: React.PropTypes.object.isRequired,
-        history: React.PropTypes.object.isRequired
+        store: React.PropTypes.shape().isRequired,
+        history: React.PropTypes.shape().isRequired
     };
 
     render() {
