@@ -7,7 +7,9 @@ export default function requireAuthentication(Component) {
 
         static propTypes = {
             isAuthenticated: React.PropTypes.bool.isRequired,
-            location: React.PropTypes.object.isRequired,
+            location: React.PropTypes.shape({
+                pathname: React.PropTypes.string.isRequired
+            }).isRequired,
             dispatch: React.PropTypes.func.isRequired
         };
 
