@@ -3,11 +3,11 @@ import os
 from django.conf import settings
 from django.http import HttpResponse
 from django.views.generic import View
+from knox.auth import TokenAuthentication
 from rest_framework import status
 from rest_framework.generics import GenericAPIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from rest_framework_jwt.authentication import JSONWebTokenAuthentication
 
 
 class IndexView(View):
@@ -21,9 +21,9 @@ class IndexView(View):
 
 
 class ProtectedDataView(GenericAPIView):
-    """Return protected data  main page."""
+    """Return protected data main page."""
 
-    authentication_classes = (JSONWebTokenAuthentication,)
+    authentication_classes = (TokenAuthentication,)
     permission_classes = (IsAuthenticated,)
 
     def get(self, request):
